perf(core): make `LOCALE_ID` and other tokens from `ApplicationModule` tree-shakable (#45102)
The `ApplicationModule` module has a number of tokens declared as non-tree-shakable providers. This commit updates them to make tree-shakable.

PR Close #45102
