+++ packages/core/src/i18n/tokens.ts
@@ -7,6 +7,39 @@
  */
 
 import {InjectionToken} from '../di/injection_token';
+import {inject} from '../di/injector_compatibility';
+import {InjectFlags} from '../di/interface/injector';
+
+import {DEFAULT_LOCALE_ID, USD_CURRENCY_CODE} from './localization';
+
+declare const $localize: {locale?: string};
+
+/**
+ * Work out the locale from the potential global properties.
+ *
+ * * Closure Compiler: use `goog.getLocale()`.
+ * * Ivy enabled: use `$localize.locale`
+ */
+export function getGlobalLocale(): string {
+  if (typeof ngI18nClosureMode !== 'undefined' && ngI18nClosureMode &&
+      typeof goog !== 'undefined' && goog.getLocale() !== 'en') {
+    // * The default `goog.getLocale()` value is `en`, while Angular used `en-US`.
+    // * In order to preserve backwards compatibility, we use Angular default value over
+    //   Closure Compiler's one.
+    return goog.getLocale();
+  } else {
+    // KEEP `typeof $localize !== 'undefined' && $localize.locale` IN SYNC WITH THE LOCALIZE
+    // COMPILE-TIME INLINER.
+    //
+    // * During compile time inlining of translations the expression will be replaced
+    //   with a string literal that is the current locale. Other forms of this expression are not
+    //   guaranteed to be replaced.
+    //
+    // * During runtime translation evaluation, the developer is required to set `$localize.locale`
+    //   if required, or just to provide their own `LOCALE_ID` provider.
+    return (typeof $localize !== 'undefined' && $localize.locale) || DEFAULT_LOCALE_ID;
+  }
+}
 
 /**
  * Provide this token to set the locale of your application.
@@ -30,7 +63,11 @@ import {InjectionToken} from '../di/injection_token';
  *
  * @publicApi
  */
-export const LOCALE_ID = new InjectionToken<string>('LocaleId');
+export const LOCALE_ID: InjectionToken<string> = new InjectionToken('LocaleId', {
+  providedIn: 'root',
+  factory: () =>
+      inject(LOCALE_ID, InjectFlags.Optional | InjectFlags.SkipSelf) || getGlobalLocale(),
+});
 
 /**
  * Provide this token to set the default currency code your application uses for
@@ -70,7 +107,10 @@ export const LOCALE_ID = new InjectionToken<string>('LocaleId');
  *
  * @publicApi
  */
-export const DEFAULT_CURRENCY_CODE = new InjectionToken<string>('DefaultCurrencyCode');
+export const DEFAULT_CURRENCY_CODE = new InjectionToken<string>('DefaultCurrencyCode', {
+  providedIn: 'root',
+  factory: () => USD_CURRENCY_CODE,
+});
 
 /**
  * Use this token at bootstrap to provide the content of your translation file (`xtb`,
