+++ packages/core/test/bundling/animations/bundle.golden_symbols.json
@@ -6,16 +6,10 @@
     "name": "ANIMATION_MODULE_TYPE"
   },
   {
-    "name": "APPLICATION_MODULE_PROVIDERS"
-  },
-  {
     "name": "APP_BOOTSTRAP_LISTENER"
   },
   {
     "name": "APP_ID"
-  },
-  {
-    "name": "APP_ID_RANDOM_PROVIDER"
   },
   {
     "name": "APP_INITIALIZER"
@@ -150,9 +144,6 @@
     "name": "DASH_CASE_REGEXP"
   },
   {
-    "name": "DEFAULT_CURRENCY_CODE"
-  },
-  {
     "name": "DEFAULT_NOOP_PREVIOUS_NODE"
   },
   {
@@ -232,9 +223,6 @@
   },
   {
     "name": "INJECTOR_SCOPE"
-  },
-  {
-    "name": "Inject"
   },
   {
     "name": "InjectFlags"
@@ -375,9 +363,6 @@
     "name": "Observable"
   },
   {
-    "name": "Optional"
-  },
-  {
     "name": "PARAM_REGEX"
   },
   {
@@ -442,9 +427,6 @@
   },
   {
     "name": "SimpleOuterSubscriber"
-  },
-  {
-    "name": "SkipSelf"
   },
   {
     "name": "SpecialCasedStyles"
@@ -616,9 +598,6 @@
   },
   {
     "name": "applyView"
-  },
-  {
-    "name": "attachInjectFlag"
   },
   {
     "name": "attachPatchData"
@@ -966,6 +945,9 @@
     "name": "initTNodeFlags"
   },
   {
+    "name": "inject"
+  },
+  {
     "name": "injectArgs"
   },
   {
@@ -1114,9 +1096,6 @@
   },
   {
     "name": "makeLambdaFromStates"
-  },
-  {
-    "name": "makeParamDecorator"
   },
   {
     "name": "makeRecord"
